test.c++: Fix "let and const heredocs" test - mistake meant heredoc continued until the end of the source text.
A line that begins with spaces cannot mark the end of a heredoc.

As the test had been with "END" indented, all the source lines after "  const k =<< trim END" would have been considered to be part of the one heredoc. (Consequently "  const k2 =<< END" would have not been interpreted as starting a new heredoc.)